+++ .github/workflows/build-complete-samples.yml
@@ -272,7 +272,7 @@ jobs:
               name: 'bot-file-upload'
               version: '6.0.x'
 
-            - project_path: 'samples/bot-initiate-thread-in-channel/csharp/TeamsStartNewThreadInTeam.csproj'
+            - project_path: 'samples/bot-initiate-thread-in-channel/csharp/TeamsStartNewThreadInTeam/TeamsStartNewThreadInTeam.csproj'
               name: 'bot-initiate-thread-in-channel'
               version: '6.0.x'
