--- samples/bot-initiate-thread-in-channel/csharp/Program.cs
@@ -1,29 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
-
-namespace Microsoft.BotBuilderSamples
-{
-    public class Program
-    {
-        public static void Main(string[] args)
-        {
-            CreateHostBuilder(args).Build().Run();
-        }
-
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-                .ConfigureWebHostDefaults(webBuilder =>
-                {
-                    webBuilder.ConfigureLogging((logging) =>
-                    {
-                        logging.AddDebug();
-                        logging.AddConsole();
-                    });
-                    webBuilder.UseStartup<Startup>();
-                });
-    }
-}
