+++ samples/bot-initiate-thread-in-channel/csharp/TeamsApp/teamsapp.local.yml
@@ -40,7 +40,7 @@ provision:
   # Generate runtime appsettings to JSON file
   - uses: file/createOrUpdateJsonFile
     with:
-      target: ../appsettings.json
+      target: ../TeamsStartNewThreadInTeam/appsettings.json
       content:
         MicrosoftAppId: ${{AAD_APP_CLIENT_ID}}
         MicrosoftAppPassword: ${{SECRET_AAD_APP_CLIENT_SECRET}}
