+++ samples/bot-initiate-thread-in-channel/csharp/TeamsStartNewThreadInTeam.sln
@@ -1,16 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.10.34814.14
+VisualStudioVersion = 17.13.35617.110
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TeamsStartNewThreadInTeam", "TeamsStartNewThreadInTeam.csproj", "{86AC8D6A-2F2E-4CB5-918A-D2D77636F9DA}"
-EndProject
 Project("{A9E3F50B-275E-4AF7-ADCE-8BE12D41E305}") = "TeamsApp", "TeamsApp\TeamsApp.ttkproj", "{8BDE6D89-0356-41BD-BE9C-F5542334F49A}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{88CDE8D8-9EEA-4318-B7F3-5821AEC9F025}"
-	ProjectSection(SolutionItems) = preProject
-		TeamsStartNewThreadInTeam.slnLaunch.user = TeamsStartNewThreadInTeam.slnLaunch.user
-	EndProjectSection
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TeamsStartNewThreadInTeam", "TeamsStartNewThreadInTeam\TeamsStartNewThreadInTeam.csproj", "{A1AE59BA-FABF-AA4F-C680-041616E6C629}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -18,19 +13,21 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{86AC8D6A-2F2E-4CB5-918A-D2D77636F9DA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{86AC8D6A-2F2E-4CB5-918A-D2D77636F9DA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{86AC8D6A-2F2E-4CB5-918A-D2D77636F9DA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{86AC8D6A-2F2E-4CB5-918A-D2D77636F9DA}.Release|Any CPU.Build.0 = Release|Any CPU
 		{8BDE6D89-0356-41BD-BE9C-F5542334F49A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{8BDE6D89-0356-41BD-BE9C-F5542334F49A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8BDE6D89-0356-41BD-BE9C-F5542334F49A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8BDE6D89-0356-41BD-BE9C-F5542334F49A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A1AE59BA-FABF-AA4F-C680-041616E6C629}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A1AE59BA-FABF-AA4F-C680-041616E6C629}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A1AE59BA-FABF-AA4F-C680-041616E6C629}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
+		{A1AE59BA-FABF-AA4F-C680-041616E6C629}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A1AE59BA-FABF-AA4F-C680-041616E6C629}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A1AE59BA-FABF-AA4F-C680-041616E6C629}.Release|Any CPU.Deploy.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {0E2F9A23-1826-4D8D-8F24-8A99E0149092}
+		SolutionGuid = {F92EE6F7-198F-45B6-B73A-6F58CB54A04D}
 	EndGlobalSection
 EndGlobal
