+++ samples/bot-initiate-thread-in-channel/csharp/TeamsStartNewThreadInTeam.slnLaunch.user
@@ -3,14 +3,14 @@
     "Name": "Microsoft Teams (browser)",
     "Projects": [
       {
-        "Path": "TeamsStartNewThreadInTeam.csproj",
-        "Action": "Start",
-        "DebugTarget": "Start Project"
-      },
-      {
         "Path": "TeamsApp\\TeamsApp.ttkproj",
         "Action": "StartWithoutDebugging",
-        "DebugTarget": "Microsoft Teams (browser)"
+        "DebugTarget": "Microsoft Teams (Browser)"
+      },
+      {
+        "Path": "TeamsStartNewThreadInTeam\\TeamsStartNewThreadInTeam.csproj",
+        "Action": "Start",
+        "DebugTarget": "Start Project"
       }
     ]
   }
