+++ samples/bot-initiate-thread-in-channel/csharp/TeamsStartNewThreadInTeam/.gitignore
