+++ samples/bot-initiate-thread-in-channel/csharp/TeamsStartNewThreadInTeam/AdapterWithErrorHandler.cs
@@ -8,23 +8,24 @@ using Microsoft.Extensions.Logging;
 
 namespace Microsoft.BotBuilderSamples
 {
+    /// <summary>
+    /// A custom adapter that includes an error handler to log and trace errors.
+    /// </summary>
     public class AdapterWithErrorHandler : CloudAdapter
     {
-        public AdapterWithErrorHandler(BotFrameworkAuthentication auth, ILogger<IBotFrameworkHttpAdapter> logger)
+        public AdapterWithErrorHandler(BotFrameworkAuthentication auth, ILogger<CloudAdapter> logger)
             : base(auth, logger)
         {
+            // Set the OnTurnError handler to log the error and trace the activity.
             OnTurnError = async (turnContext, exception) =>
             {
-                // Log any leaked exception from the application.
-                // NOTE: In production environment, you should consider logging this to
-                // Azure Application Insights. Visit https://aka.ms/bottelemetry to see how
-                // to add telemetry capture to your bot.
+                // Log any exceptions to the configured logger.
                 logger.LogError($"Exception caught : {exception.Message}");
 
                 // Uncomment below commented line for local debugging..
                 // await turnContext.SendActivityAsync($"Sorry, it looks like something went wrong. Exception Caught: {exception.Message}");
 
-                // Send a trace activity, which will be displayed in the Bot Framework Emulator
+                // Send a trace activity, which will be displayed in the Bot Framework Emulator.
                 await turnContext.TraceActivityAsync("OnTurnError Trace", exception.Message, "https://www.botframework.com/schemas/error", "TurnError");
             };
         }
