+++ samples/bot-initiate-thread-in-channel/csharp/TeamsStartNewThreadInTeam/DeploymentTemplates/DeployUseExistResourceGroup/parameters-for-template-AzureBot-with-rg.json
