+++ samples/bot-initiate-thread-in-channel/csharp/TeamsStartNewThreadInTeam/DeploymentTemplates/DeployUseExistResourceGroup/parameters-for-template-BotApp-with-rg.json
