+++ samples/bot-initiate-thread-in-channel/csharp/TeamsStartNewThreadInTeam/DeploymentTemplates/DeployUseExistResourceGroup/template-AzureBot-with-rg.json
