+++ samples/bot-initiate-thread-in-channel/csharp/TeamsStartNewThreadInTeam/DeploymentTemplates/DeployUseExistResourceGroup/template-BotApp-with-rg.json
