+++ samples/bot-initiate-thread-in-channel/csharp/TeamsStartNewThreadInTeam/DeploymentTemplates/DeployWithNewResourceGroup/template-AzureBot-new-rg.json
