+++ samples/bot-initiate-thread-in-channel/csharp/TeamsStartNewThreadInTeam/DeploymentTemplates/DeployWithNewResourceGroup/template-BotApp-new-rg.json
