+++ samples/bot-initiate-thread-in-channel/csharp/TeamsStartNewThreadInTeam/Properties/launchSettings.json
