+++ samples/bot-initiate-thread-in-channel/csharp/TeamsStartNewThreadInTeam/appsettings.json
@@ -1,0 +1,6 @@
+{
+	"MicrosoftAppType": "",
+	"MicrosoftAppId": "",
+	"MicrosoftAppPassword": "",
+	"MicrosoftAppTenantId": ""
+}
