+++ samples/bot-initiate-thread-in-channel/csharp/TeamsStartNewThreadInTeam/wwwroot/default.html
