--- samples/bot-initiate-thread-in-channel/csharp/appsettings.json
@@ -1,6 +1,0 @@
-{
-  "MicrosoftAppType": "",
-  "MicrosoftAppId": "",
-  "MicrosoftAppPassword": "",
-  "MicrosoftAppTenantId": ""
-}
