--- samples/bot-initiate-thread-in-channel/csharp/infra/azure.bicep
@@ -1,44 +1,0 @@
-@maxLength(20)
-@minLength(4)
-@description('Used to generate names for all resources in this file')
-param resourceBaseName string
-
-@description('Required when create Azure Bot service')
-param botAadAppClientId string
-
-param botAppDomain string
-
-@maxLength(42)
-param botDisplayName string
-
-param botServiceName string = resourceBaseName
-param botServiceSku string = 'F0'
-param microsoftAppType string
-param microsoftAppTenantId string
-
-// Register your web service as a bot with the Bot Framework
-resource botService 'Microsoft.BotService/botServices@2021-03-01' = {
-  kind: 'azurebot'
-  location: 'global'
-  name: botServiceName
-  properties: {
-    displayName: botDisplayName
-    endpoint: 'https://${botAppDomain}/api/messages'
-    msaAppId: botAadAppClientId
-    msaAppType: microsoftAppType
-    msaAppTenantId: microsoftAppType == 'SingleTenant' ? microsoftAppTenantId : ''
-  }
-  sku: {
-    name: botServiceSku
-  }
-}
-
-// Connect the bot service to Microsoft Teams
-resource botServiceMsTeamsChannel 'Microsoft.BotService/botServices/channels@2021-03-01' = {
-  parent: botService
-  location: 'global'
-  name: 'MsTeamsChannel'
-  properties: {
-    channelName: 'MsTeamsChannel'
-  }
-}
