--- samples/bot-initiate-thread-in-channel/csharp/infra/azure.parameters.json
@@ -1,24 +1,0 @@
-{
-  "$schema": "https://schema.management.azure.com/schemas/2015-01-01/deploymentParameters.json#",
-  "contentVersion": "1.0.0.0",
-  "parameters": {
-    "resourceBaseName": {
-      "value": "bot${{RESOURCE_SUFFIX}}"
-    },
-    "botAadAppClientId": {
-      "value": "${{AAD_APP_CLIENT_ID}}"
-    },
-    "botAppDomain": {
-      "value": "${{BOT_DOMAIN}}"
-    },
-    "botDisplayName": {
-      "value": "bot-initiate-thread-in-channel"
-    },
-    "microsoftAppType": {
-      "value": "${{MICROSOFT_APP_TYPE}}"
-    },
-    "microsoftAppTenantId": {
-      "value": "${{MICROSOFT_APP_TENANT_ID}}"
-    }
-  }
-}
