--- samples/bot-initiate-thread-in-channel/csharp/teamsapp.yml
@@ -1,9 +1,0 @@
-# yaml-language-server: $schema=https://aka.ms/teams-toolkit/v1.2/yaml.schema.json
-# Visit https://aka.ms/teamsfx-v5.0-guide for details on this file
-# Visit https://aka.ms/teamsfx-actions for details on actions
-version: v1.2
-
-additionalMetadata:
-  sampleTag: Microsoft-Teams-Samples:bot-initiate-thread-in-channel-csharp
-
-environmentFolderPath: ./env
